docs/peps/001-instrumentation-tempo/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Grafana Tempo poc app"
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
opentelemetry-instrumentation = "^0.49b2"
opentelemetry-distro = "^0.49b2"
opentelemetry-exporter-otlp = "^1.28.2"
opentelemetry-instrumentation-asyncio = "^0.49b2"
opentelemetry-instrumentation-threading = "^0.49b2"
opentelemetry-instrumentation-dbapi = "^0.49b2"
opentelemetry-instrumentation-logging = "^0.49b2"
opentelemetry-instrumentation-urllib = "^0.49b2"
opentelemetry-instrumentation-wsgi = "^0.49b2"
opentelemetry-instrumentation-grpc = "^0.49b2"
opentelemetry-instrumentation-requests = "^0.49b2"
opentelemetry-instrumentation-urllib3 = "^0.49b2"
opentelemetry-api = "^1.28.2"
opentelemetry-processor-baggage = "^0.49b2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
